refactor(struct): migrate core consumers to package exports
