+++ package.json
@@ -81,6 +81,7 @@
     "@astrojs/react": "^6.0.1",
     "@astrojs/rss": "^4.0.19",
     "@astrojs/sitemap": "^3.7.3",
+    "@erniesg/struct": "file:packages/struct",
     "@hbsnow/rehype-sectionize": "^1.0.7",
     "@iconify-json/lucide": "^1.2.20",
     "@mdx-js/mdx": "3.1.1",
